test(ui): split catalog request-error cases under max-lines
