+++ ui/src/test-helpers/app-sidebar-cases/catalog-live.ts
@@ -65,47 +65,6 @@ describe("AppSidebar session catalog pagination", () => {
     }
   });
 
-  it("does not retry unrelated INVALID_REQUEST failures and renders a local retry", async () => {
-    vi.useFakeTimers();
-    const warn = vi.spyOn(console, "warn").mockImplementation(() => undefined);
-    try {
-      const request = vi.fn().mockRejectedValue(
-        new GatewayRequestError({
-          code: "INVALID_REQUEST",
-          message: 'unknown agent id "main"',
-        }),
-      );
-      const gateway = createGatewayHarness({ request } as unknown as GatewayBrowserClient);
-      gateway.publish({
-        hello: {
-          features: { methods: ["sessions.catalog.list"] },
-        } as ApplicationGatewaySnapshot["hello"],
-      });
-      const { sidebar } = await mountSidebar(
-        gateway.gateway,
-        createSessions("main", ["agent:main:main"]),
-      );
-      sidebar.connected = true;
-      await sidebar.updateComplete;
-      await vi.advanceTimersByTimeAsync(0);
-      await sidebar.updateComplete;
-
-      expect(request).toHaveBeenCalledTimes(1);
-      const error = sidebar.querySelector<HTMLElement>(".sidebar-session-catalog-error");
-      expect(error?.textContent).toContain('unknown agent id "main"');
-
-      error?.querySelector<HTMLButtonElement>("button")?.click();
-      await vi.advanceTimersByTimeAsync(0);
-      await sidebar.updateComplete;
-
-      expect(request).toHaveBeenCalledTimes(2);
-      expect(warn).toHaveBeenCalledTimes(1);
-    } finally {
-      warn.mockRestore();
-      vi.useRealTimers();
-    }
-  });
-
   it("ignores a legacy fallback that settles after the Gateway client changes", async () => {
     vi.useFakeTimers();
     try {
